+++ android/assets/tilesprites/3.txt
@@ -1,0 +1,116 @@
+3.png
+format: RGBA8888
+filter: Nearest,Nearest
+repeat: none
+1
+  rotate: false
+  xy: 2, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+10
+  rotate: false
+  xy: 2, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+11
+  rotate: false
+  xy: 5, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+12
+  rotate: false
+  xy: 5, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+13
+  rotate: false
+  xy: 8, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+14
+  rotate: false
+  xy: 8, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+15
+  rotate: false
+  xy: 11, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+16
+  rotate: false
+  xy: 11, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+2
+  rotate: false
+  xy: 14, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+3
+  rotate: false
+  xy: 14, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+4
+  rotate: false
+  xy: 17, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+5
+  rotate: false
+  xy: 17, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+6
+  rotate: false
+  xy: 20, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+7
+  rotate: false
+  xy: 20, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+8
+  rotate: false
+  xy: 23, 2
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
+9
+  rotate: false
+  xy: 23, 5
+  size: 1, 1
+  orig: 1, 1
+  offset: 0, 0
+  index: -1
