+++ android/src/pl/gratitude/brighter/AndroidLauncher.java
@@ -1,0 +1,19 @@
+package pl.gratitude.brighter;
+
+import android.os.Bundle;
+
+import com.badlogic.gdx.backends.android.AndroidApplication;
+import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
+
+public class AndroidLauncher extends AndroidApplication {
+	@Override
+	protected void onCreate (Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		AndroidApplicationConfiguration config = new AndroidApplicationConfiguration();
+		config.useAccelerometer = false;
+		config.useCompass = false;
+		config.useImmersiveMode = false;
+		config.hideStatusBar = false;
+		initialize(new Main(), config);
+	}
+}
