+++ core/src/pl/gratitude/brighter/handlers/AssetsHandler.java
@@ -1,0 +1,9 @@
+package pl.gratitude.brighter.handlers;
+
+/**
+ * Created on 23.02.2016
+ *
+ * @author Sławomir Onyszko
+ */
+public class AssetsHandler {
+}
