+++ core/src/pl/gratitude/brighter/states/MenuState.java
@@ -1,0 +1,35 @@
+package pl.gratitude.brighter.states;
+
+import pl.gratitude.brighter.utils.GameStateManager;
+
+/**
+ * Created on 23.02.2016
+ *
+ * @author Sławomir Onyszko
+ */
+public class MenuState extends BaseState {
+
+    public MenuState(GameStateManager gsm) {
+        super(gsm);
+    }
+
+    @Override
+    public void resize(int width, int height) {
+        super.resize(width, height);
+    }
+
+    @Override
+    public void update(float dt) {
+        super.update(dt);
+    }
+
+    @Override
+    public void render() {
+        super.render();
+    }
+
+    @Override
+    public void dispose() {
+        super.dispose();
+    }
+}
