--- core/src/pl/onyszko/brighter/handlers/AssetsHandler.java
@@ -1,9 +1,0 @@
-package pl.onyszko.brighter.handlers;
-
-/**
- * Created on 23.02.2016
- *
- * @author Sławomir Onyszko
- */
-public class AssetsHandler {
-}
