--- core/src/pl/onyszko/brighter/interfaces/StateInterface.java
@@ -1,15 +1,0 @@
-package pl.onyszko.brighter.interfaces;
-
-/**
- * Created on 23.02.2016
- *
- * @author Sławomir Onyszko
- */
-public interface StateInterface {
-
-    void resize(int width, int height);
-    void update(float dt);
-    void render();
-    void dispose();
-
-}
