--- core/src/pl/onyszko/brighter/utils/Dictionary.java
@@ -1,15 +1,0 @@
-package pl.onyszko.brighter.utils;
-
-/**
- * Created on 23.02.2016
- *
- * @author Sławomir Onyszko
- */
-public class Dictionary {
-
-    public static final String GAME_TITLE = "Brighter";
-    public static final int VIRTUAL_WIDTH = 960;
-    public static final int VIRTUAL_HEIGHT = 800;
-    public static final int TILE_SPACING = 8;
-
-}
