+++ ios/robovm.properties
@@ -1,6 +1,6 @@
 app.version=1.0
-app.id=pl.onyszko.brighter.IOSLauncher
-app.mainclass=pl.onyszko.brighter.IOSLauncher
+app.id=pl.gratitude.brighter.IOSLauncher
+app.mainclass=pl.gratitude.brighter.IOSLauncher
 app.executable=IOSLauncher
 app.build=1
 app.name=Brighter
